rewrote RS232 driver model: jump table instead of KERNAL hooks (#7)
CCGMS Term has historically used the KERNAL interface to the device 2 user port ACIA 6551 emulation and later hooked KERNAL vectors to provide an improved driver. As of CCGMS 2021, the three drivers (User Port, UP9600 and SwiftLink) each hook IGETIN and IBSOUT so the existing code can continue to use BSOUT/GETIN.

Apart from the unnecessary indirection and the extra code needed, this has the side effect of still invoking parts of the KERNAL RS232 driver on OPEN and CHKIN, which writes to CIA2 registers and may cause incompatibilities with some hardware. 

(In fact, every CHKIN clears FLAG, causing the WiC64 hardware to miss the transmission of a byte.)

Some code already used `modget` to get a byte from the modem. This change moves all RS232 accesses to a dedicated jump-table-based interface.

Some details:

* The user port driver relies on some setup by KERNAL code, so an explicit call to $f404 was added in one code path.
* RS232 dispatch was cleaned up and moved into rs232.s
* SwiftLink address patching was moved into the SwiftLink driver.